% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCSO.R
\docType{data}
\name{CCSO}
\alias{CCSO}
\title{Champaign County Sheriff's Office Booking Data}
\format{
A tibble with 67764 observations of 31 variables:
}
\source{
https://data.illinois.gov/dataset/jail-booking-data
}
\usage{
data(CCSO)
}
\description{
Arrest data obtained from the Champaign County Sheriff's Office
located in Champaign, IL Data collected from Jan 1, 2011 through
December 31, 2016.
}
\keyword{datasets}
